// plugins/plugin-s3/ibm/src/model/ServiceKey.ts
/*
 * Copyright 2021 The Kubernetes Authors
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import Common from './Common'

type ServiceKey = Common & {
  source_crn: string
  credentials: {
    apikey: string
    cos_hmac_keys?: {
      access_key_id: string
      secret_access_key: string
    }
    endpoints: string
    iam_apikey_description: string
    iam_apikey_name: string
    iam_role_crn: string
    iam_serviceid_crn: string
    resource_instance_id: string
  }
  iam_compatible: boolean
  resource_instance_url: string
}

export default ServiceKey